+++ models/issue_stopwatch_test.go
@@ -3,7 +3,7 @@ package models
 import (
 	"testing"
 
-	"code.gitea.io/gitea/modules/util"
+	"code.gitea.io/gitea/modules/timeutil"
 
 	"github.com/stretchr/testify/assert"
 )
